Application: Common.h : changed cast to static_cast FilterWorker: - added pids in while loop, we collect pids we handle irps, and then decide if malicious foreach pid - clean code - added code to check malicious pid and base code to handle malicious apps. Globals: - added num of files in protected directories. - added num of subdirs in protected directories. - added setters and getters HashUtility: Removed MyForm: - clean code - removed button initTraps - addDirectory and remDirectory now add/remove directories for traps generations Process: - added check if fileid is traps in irps handling - allow extensions save - base code for self check if malicious Thresholds: base values added in header TrapHandler: - removed map of traps (not used) - clean code - only addDirTraps and remDirTraps are now public, those work recursivly on a given string they receive which should be a directory path. - code of addDir and remDir changed to managed code which uses TrapsMemory singleton
Driver:
- Buffer of file paths is now bigger 512 WCHAR letters
- added memory deallocation which was missing
- added copy for the files extensions in case of read/write irp

SharedDefs:
Increase of MAX_FILE_NAME_LENGTH buffer

+++ Application/Common.h
@@ -65,8 +65,8 @@ value struct FileId {
 
 	virtual int GetHashCode() override {
 		//interior_ptr<ULONG> ptr = &volumeSerialNumber;
-		ULONG val1 = volumeSerialNumber;
-		ULONG val2 = (volumeSerialNumber >> (sizeof(ULONG) * 8));
+		ULONG val1 = static_cast<ULONG> (volumeSerialNumber);
+		ULONG val2 = static_cast<ULONG>((volumeSerialNumber >> (sizeof(ULONG) * 8)));
 		ULONG val3 = (((ULONG)(fileId[0])) | (((ULONG)(fileId[1])) << 8) | (((ULONG)(fileId[2])) << 16) | (((ULONG)(fileId[3])) << 24));
 		ULONG val4 = (((ULONG)(fileId[4])) | (((ULONG)(fileId[5])) << 8) | (((ULONG)(fileId[6])) << 16) | (((ULONG)(fileId[7])) << 24));
 		ULONG val5 = (((ULONG)(fileId[8])) | (((ULONG)(fileId[9])) << 8) | (((ULONG)(fileId[10])) << 16) | (((ULONG)(fileId[11])) << 24));
